refactor: using react-native-compat to detect installed apps
