--- android/gradle.properties
@@ -1,5 +1,0 @@
-InstalledApp_kotlinVersion=1.7.0
-InstalledApp_minSdkVersion=21
-InstalledApp_targetSdkVersion=31
-InstalledApp_compileSdkVersion=31
-InstalledApp_ndkversion=21.4.7075529
