--- android/src/newarch/InstalledAppSpec.kt
@@ -1,7 +1,0 @@
-package com.installedapp
-
-import com.facebook.react.bridge.ReactApplicationContext
-
-abstract class InstalledAppSpec internal constructor(context: ReactApplicationContext) :
-  NativeInstalledAppSpec(context) {
-}
