+++ package.json
@@ -9,17 +9,6 @@
   "files": [
     "src",
     "lib",
-    "android",
-    "ios",
-    "cpp",
-    "*.podspec",
-    "!lib/typescript/example",
-    "!ios/build",
-    "!android/build",
-    "!android/gradle",
-    "!android/gradlew",
-    "!android/gradlew.bat",
-    "!android/local.properties",
     "!**/__tests__",
     "!**/__fixtures__",
     "!**/__mocks__",
@@ -67,6 +56,7 @@
     "@types/qrcode": "1.5.0",
     "@types/react": "17.0.21",
     "@types/react-native": "0.70.0",
+    "@walletconnect/react-native-compat": "2.10.5",
     "@walletconnect/types": "2.5.2",
     "commitlint": "17.5.1",
     "del-cli": "5.0.0",
@@ -92,6 +82,7 @@
   "peerDependencies": {
     "@react-native-async-storage/async-storage": ">=1.17.0",
     "@react-native-community/netinfo": ">=9.0.0",
+    "@walletconnect/react-native-compat": ">=2.10.5",
     "react": "*",
     "react-native": "*",
     "react-native-get-random-values": ">=1.8.0",
@@ -146,9 +137,8 @@
     ]
   },
   "dependencies": {
-    "@walletconnect/core": "2.10.4",
-    "@walletconnect/react-native-compat": "2.10.4",
-    "@walletconnect/universal-provider": "2.10.4",
+    "@walletconnect/core": "2.10.5",
+    "@walletconnect/universal-provider": "2.10.5",
     "qrcode": "1.5.3",
     "valtio": "1.10.5"
   }
