+++ src/utils/ExplorerUtil.ts
@@ -8,8 +8,6 @@ import type {
 import { CoreUtil } from './CoreUtil';
 import { ConfigCtrl } from '../controllers/ConfigCtrl';
 import { ToastCtrl } from '../controllers/ToastCtrl';
-import { isAppInstalled } from '../modules/AppInstalled';
-import { PLAYSTORE_REGEX } from '../constants/Platform';
 import { StorageUtil } from './StorageUtil';
 import { CORE_VERSION, SDK_VERSION } from '../constants/Config';
 
@@ -45,31 +43,6 @@ async function fetchListings(
   const request = await fetch(url.toString(), { headers });
 
   return request.json() as Promise<ListingResponse>;
-}
-
-function getUrlParams(url: string | null): { [key: string]: string } {
-  if (!url) {
-    return {};
-  }
-  const regex = /[?&]([^=#]+)=([^&#]*)/g;
-  const params: { [key: string]: string } = {};
-  let match: RegExpExecArray | null;
-
-  while ((match = regex.exec(url)) !== null) {
-    params[match[1] as string] = decodeURIComponent(match[2] as string);
-  }
-
-  return params;
-}
-
-function getAppId(playstoreLink?: string | null): string | undefined {
-  if (!playstoreLink || !playstoreLink.match(PLAYSTORE_REGEX)) {
-    return undefined;
-  }
-
-  // eslint-disable-next-line dot-notation
-  const applicationId = getUrlParams(playstoreLink)?.['id'];
-  return applicationId;
 }
 
 // -- Utility -------------------------------------------------------
@@ -138,23 +111,11 @@ export const ExplorerUtil = {
     };
   },
 
-  async isAppInstalled(wallet: Listing): Promise<boolean> {
-    let isInstalled = false;
-    const scheme = wallet.mobile.native;
-    const appId = getAppId(wallet.app.android);
-    try {
-      isInstalled = await isAppInstalled(scheme, appId);
-    } catch {
-      isInstalled = false;
-    }
-    return isInstalled;
-  },
-
   async sortInstalled(array: Listing[]) {
     const promises = array.map(async (item) => {
       return {
         ...item,
-        isInstalled: await ExplorerUtil.isAppInstalled(item),
+        isInstalled: await CoreUtil.checkInstalled(item),
       };
     });
